MVP `no_std` support in `wgpu-hal`
