+++ Cargo.toml
@@ -74,7 +74,7 @@ wgpu-core = { version = "25.0.0", path = "./wgpu-core" }
 wgpu-hal = { version = "25.0.0", path = "./wgpu-hal" }
 wgpu-macros = { version = "25.0.0", path = "./wgpu-macros" }
 wgpu-test = { version = "25.0.0", path = "./tests" }
-wgpu-types = { version = "25.0.0", path = "./wgpu-types" }
+wgpu-types = { version = "25.0.0", path = "./wgpu-types", default-features = false }
 
 wgpu-core-deps-windows-linux-android = { version = "25.0.0", path = "./wgpu-core/platform-deps/windows-linux-android" }
 wgpu-core-deps-apple = { version = "25.0.0", path = "./wgpu-core/platform-deps/apple" }
