+++ wgpu-hal/src/lib.rs
@@ -276,6 +276,12 @@ pub mod api {
 
 mod dynamic;
 
+#[cfg(feature = "validation_canary")]
+mod validation_canary;
+
+#[cfg(feature = "validation_canary")]
+pub use validation_canary::{ValidationCanary, VALIDATION_CANARY};
+
 pub(crate) use dynamic::impl_dyn_resource;
 pub use dynamic::{
     DynAccelerationStructure, DynAcquiredSurfaceTexture, DynAdapter, DynBindGroup,
@@ -298,7 +304,6 @@ use core::{
 };
 
 use bitflags::bitflags;
-use parking_lot::Mutex;
 use thiserror::Error;
 use wgt::WasmNotSendSync;
 
@@ -2372,42 +2377,6 @@ pub struct RenderPassDescriptor<'a, Q: DynQuerySet + ?Sized, T: DynTextureView +
 pub struct ComputePassDescriptor<'a, Q: DynQuerySet + ?Sized> {
     pub label: Label<'a>,
     pub timestamp_writes: Option<PassTimestampWrites<'a, Q>>,
-}
-
-/// Stores the text of any validation errors that have occurred since
-/// the last call to `get_and_reset`.
-///
-/// Each value is a validation error and a message associated with it,
-/// or `None` if the error has no message from the api.
-///
-/// This is used for internal wgpu testing only and _must not_ be used
-/// as a way to check for errors.
-///
-/// This works as a static because `cargo nextest` runs all of our
-/// tests in separate processes, so each test gets its own canary.
-///
-/// This prevents the issue of one validation error terminating the
-/// entire process.
-pub static VALIDATION_CANARY: ValidationCanary = ValidationCanary {
-    inner: Mutex::new(Vec::new()),
-};
-
-/// Flag for internal testing.
-pub struct ValidationCanary {
-    inner: Mutex<Vec<String>>,
-}
-
-impl ValidationCanary {
-    #[allow(dead_code)] // in some configurations this function is dead
-    fn add(&self, msg: String) {
-        self.inner.lock().push(msg);
-    }
-
-    /// Returns any API validation errors that have occurred in this process
-    /// since the last call to this function.
-    pub fn get_and_reset(&self) -> Vec<String> {
-        self.inner.lock().drain(..).collect()
-    }
 }
 
 #[test]
